+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Uphold parser: new data file format added.
 - Poloniex parser: new "trades" data file format added.
 - Poloniex parser: new "distributions" data file format added.
+- Conversion tool: added colour bands to Excel output file.
 ### Changed
 - Sort wallet names in audit debug as case-insensitive.
 - Data source names in config are now case-insensitive.